Move the Requirement API to the package root
